pyproject.toml
[tool.poetry]
name = "nonebot_plugin_tuan_chatgpt"
version = "0.4.2"
description = "\"Chat with tuanzi ~\""
authors = ["TheLZY"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/TheLZY/nonebot_plugin_tuan_chatgpt"

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.27.0"
tiktoken = "^0.3.0"
nonebot2 = "^2.0.0rc3"
nonebot-adapter-onebot = "^2.2.0"
aiohttp = "^3.8.4"
urllib3 = "1.25.11"
Pillow = "^9.5.0"
qrcode = "^7.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
